decoupled config, wrapped with cli
