+++ config.js
@@ -1,0 +1,10 @@
+const config = {
+    saimeUrl: 'https://tramites.saime.gob.ve/index.php?r=usuario/usuario/BuscarSaimeContacto',
+    cneUrlBase: 'http://www.cne.gov.ve/web/registro_electoral/ce.php',
+    mongoUrl: 'mongodb://localhost:27017',
+    saimeCollection: 'saime',
+    cneCollection: 'cne',
+    mongoDB: 'venez',
+}
+
+module.exports = config;
